+++ src/main/java/nl/esciencecenter/xenon/cli/listjobs/ListJobsOutput.java
@@ -3,42 +3,61 @@ package nl.esciencecenter.xenon.cli.listjobs;
 import java.util.List;
 import java.util.Objects;
 
+import nl.esciencecenter.xenon.schedulers.JobStatus;
+
 /**
  * Listing of jobs
  */
 public class ListJobsOutput {
-    private final String location;
-    private final String queue;
-    public final List<String> jobs;
+    public final List<JobStatus> statuses;
 
-    ListJobsOutput(String location, String queue, List<String> jobIdentifiers) {
-        this.location = location;
-        this.queue = queue;
-        this.jobs = jobIdentifiers;
+
+    public ListJobsOutput(List<JobStatus> statuses) {
+        this.statuses = statuses;
     }
 
     @Override
     public String toString() {
-        String sep = System.getProperty("line.separator");
-        return String.join(sep, jobs) + sep;
+        StringBuilder sb = new StringBuilder();
+        String sep = "\t";
+        String lsep = System.getProperty("line.separator");
+        sb.append(String.join(
+            sep,
+            "Job identifier",
+            "State", "Running",
+            "Done", "Error",
+            "Exit code",
+            "Information"
+        )).append(lsep);
+        for (JobStatus status: statuses) {
+            sb.append(String.join(
+                sep,
+                status.getJobIdentifier(),
+                status.getState(),
+                String.valueOf(status.isRunning()),
+                String.valueOf(status.isDone()),
+                status.hasException() ? status.getException().getMessage() : "",
+                status.getExitCode() != null ? status.getExitCode().toString() : "",
+                status.getSchedulerSpecficInformation() != null ? status.getSchedulerSpecficInformation().toString() : ""
+            )).append(lsep);
+        }
+        return sb.toString();
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) {
             return true;
-        }    
+        }
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
         ListJobsOutput that = (ListJobsOutput) o;
-        return Objects.equals(location, that.location) &&
-                Objects.equals(queue, that.queue) &&
-                Objects.equals(jobs, that.jobs);
+        return Objects.equals(statuses, that.statuses);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(location, queue, jobs);
+        return Objects.hash(statuses);
     }
 }