+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* Status details to jobs list
 * --verbose and --stacktrace arguments
 * In `xenon --help`, added type column to list of adaptors